requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
six
django-redis
celery[redis]
django-redis
python-dotenv
drf-spectacular